# Cleaned up auto decoding special case in std.conv.parse for integers