Based on Andrei's advice in https://github.com/dlang/phobos/pull/4674, I cleaned up the code here.

Now even faster. 

Built with `ldc2 -O5 -release`, [code here](https://gist.github.com/JackStouffer/42aae01b9cd4c86ae42086642270b407#file-test-d).

```
# new
Total time: 10 ms, 730 μs, and 5 hnsecs

#current
Total time: 36 ms, 573 μs, and 4 hnsecs

#old
Total time: 181 ms, 850 μs, and 5 hnsecs
```