@JackStouffer 2017-01-04
Auto-merge toggled on